syscall: change stack alignment trick
The stack alignment trick I used in syscall_entry() worked, but the
CFI code was actually wrong and only worked in for certain alignments
but not others.

To get gdb to work in either stack alignment cases, I needed to move the
place of the alignment trick from the end, to the beginning - so when
gdb sees %rsp at the end of all the pushes, it can calculate back where
we stored all the registers, without caring that we moved all the stack
(or not) in the beginning.

This is really funky, but it appears to work (both functionally, and in
gdb).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>