syscall: fix (again) gdb support
The CFI code in syscall_entry still was not perfect (sorry...) - when
the application uses -fomit-frame-pointer, and %rbp was not set by the
caller, we didn't tell CFI how to restore the caller's %rsp. We need to
do this, with some more trickery :-(

Backtracing a crash in a system call in gdb should now be possible
(I hope) in cases which failed previously.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>