sched: only allocate sched::thread objects on the heap
In issue #790, Timmons C. Player discovered something dangerous with
on-stack sched::thread objects: The application's stack may be mmap()ed,
so the scheduler now needs to access an application's mmap()ed memory
area. Those accesses can potentially have all sort of problems (e.g.,
page faults in the scheduler, although in practice issue #143 precludes
it). But more concretely, The "lazy TLB flush" code added in commit
7e38453 means that the scheduler may see old pages for an application's
mmap()ed pages, so it cannot work with a sched::thread in an mmap()ed
area.

This patch prevents on-stack sched::thread construction by hiding the
constructor, and only allowing a thread to be created through the function
sched::thread::make(...), which creates the object on the heap.

This patch is long but it more-or-less contains just the following
types of changes:

1. The change to sched.hh to hide the sched::thread constructor, and
   instead offer a sched::thread::make(...) function.

2. Every class which used a "sched::thread" member was replaced by
   a std::unique_ptr<sched::thread> member.

   One "casualty" of this change is that a class that used to hold a
   sched::thread member will now hold a sched::thread pointer, and use an
   extra allocation and indirection - even if we know for sure that the
   containing object will always be allocated on the heap (such as the
   case in pthread.cc's pthread::_thread, for example). This can be worked
   around by making the exceptional class a friend of sched::thread to be
   able to use the hidden constructor - but for now I decided the tiny
   performance benefit isn't important enough to make this exception.

3. Every place which used "new sched::thread(...)" to create the thread,
   now uses "sched::thread::make(...)" instead.

   Sadly we no longer allow "new sched::thread(...)", although there is
   nothing wrong with it, because disabling the constructor also disables
   the new expression.

4. One test in misc-wake.cc (which wasn't run by default anyway, because)
   it is named "misc-*") was commented out because it relied on creating
   thread objects in mmap()ed areas.

One of the places modified is rcu_flush(), whose on-stack sched::thread
objects are what caused issue #790.

Fixes #790.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>