cloud-init: use hostname from yaml file
Previously cloud-init set VM hostname if EC2/GCE was used as source. But
it ignored hostname if it was specified  via yaml file or
contextualization image. The commit fixes this, so that hostname can be
set via yaml file too.

The example cloud-init.yaml file is updated with commented-out examples
for setting hostname and NFS mounts.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Message-Id: <1479395213-14565-1-git-send-email-justin.cinkelj@xlab.si>