command line: run simple scripts
runcript option is added to command line parser to allow running
simple one-line scripts. Say VM image has file named "/zpool-list" with
content "/zpool.so list". Then
run.py -e "runscript /zpool-list; /cli/cli.so"
is equivalent to
run.py -e "/zpool.so list; /cli/cli.so"

Usecase of this patch is for example HDFS. There we have one java
program, which can be run as (one) namenode or as (multiple) datanodes.
We whish to provide convinient shortcut for selection of the two distinc
java commandline parameters. So we put all required parameters into
two script-like files, and user only has to select one of them.

The need for this arose from using OSv on openstack. We build image
with capstan, and at build time we do not know what purpose it will serve
(e.g namenode or datanode). With pre-prepared run configuration/script
we ensure correct mandatory parameters are supplied (and that they
are in sync between namenode and datanodes).

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Message-Id: <1479825169-15499-2-git-send-email-justin.cinkelj@xlab.si>