dhcp: send DHCP release at shutdown
The release packet needs to contain client and server IP, so they are
stored to dhcp_interface_state when DHCP ACK is received.
During shutdown, each interface with DHCP address releases its DHCP IP.

The code does not properly wait for UDP packet to be actually put
on wire. Just a simple usleep is used to hopefully wait long enough.
See also #557 and #720.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Message-Id: <1479829966-26491-1-git-send-email-justin.cinkelj@xlab.si>