malloc: make it easier to debug ridiculous allocations
As explained in issue #585, if an application's malloc() attempt fails
because of lack of memory, we abort inside the "reclaimer thread", making
it hard to debug where in the application the OOM actually happened.

This patch doesn't solve the whole problem, but solves an important
special case: If the application tries to allocate a ridiculous amount,
larger than the entire VM memory, with this patch the abort will happen
during the allocation attempt, making it easy to debug where the
application did the wrongly-sized malloc().

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1480001709-673-1-git-send-email-nyh@scylladb.com>