build: strip binaries in modules/ dir as well
upload_manifest.py stripped only binaries in the build/ directory, so
binaries in modules, such as modules/httpserver/libhttpserver.so were
not stripped.

Arguably, each module's Makefile could strip its own results, but it's
convenient to keep both the unstripped binary (for debugging) and stripped
one (for the image). So this patch extends to modules/ and apps/ the
already existing mechanism which when uploading a file something.so,
strips it into something-stripped.so and uploads that.

The patch also ensures that even if something-stripped.so matches a
wildcard it is not independently upload to the image (causing a loop).

With this patch, the image generated by "scripts/build image=httpserver"
is reduced from 25 MB to 17 MB.

Refs #822

Signed-off-by: Yuri Volchkov <iurii.volchkov@neclab.eu>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>