signal: remove "nested signals" protection
This reverts commit 4af3677117a5bb2f4660d750fa4eddc6ef44e7f2 which
aborted on a "nested signal" (handling a signal handler while inside
a signal handler).

Nested signals were never strictly impossible, but even more importantly,
if a signal handler is exited with a longjmp or siglongjmp, our test code
would think it is still inside the handler, and the next signal handler
called by thi thread will abort on a supposedly "nested signal".

Perhaps this test can be redeemed if siglongjmp() would also clear the
signal_nesting flag, but it's easier just to remove this test.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170110013215.25950-1-nyh@scylladb.com>