mount: fix deadlock
This patch fixes a deadlock caused by lock order inversion. This deadlock,
described in issue #817, sometimes hangs the entire VFS when a mount is
attempted.

Most of the VFS code takes locks on various nodes, and later, in internal
ZFS code, calls vfs_busy() which takes the mount_lock.
Before this patch, the sys_mount() code took these locks in the opposite
order: mount_lock first, and while this is held, then called lookup code
which locked various nodes.

The fix is for sys_mount() to not hold mount_lock while looking up paths,
and only hold it while actually reading or writing the mount_list object.
Because we do this twice in sys_mount - in the beginning to check if the
path is not already mounted, and at the end to add the newly prepared
mount point to the list - the patch also adds a new mutex to serialize
calls to sys_mount() so that we won't have a situation where two concurrent
mounts both decide they can mount the same device or on the same path.

Fixes #817.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170113133550.25577-2-nyh@scylladb.com>