sched::thread: add unpin() method
This patch adds an unpin() method to undo a pin() operation on a thread
(operating on either the current thread or on a different thread is
supported). After an unpin() operation, the thread may be migrated
to any CPU (unless, of course, a temporary migrate_disable() is in force).

This patch does not yet change sched_setaffinity() and friends to call
the new unpin() method. This will come in the next patch.

Refs #778.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170117135735.19891-1-nyh@scylladb.com>