<osv/app.hh> - remove _KERNEL restriction
In <osv/app.hh>, all of the osv::application:: API was hidden with a
'#ifdef _KERNEL', probably to avoid header file conflicts in applications
which include the normal Linux header files. I fixed these conflicts in
the previous patch, so we no longer need this _KERNEL stuff. It means
that an application which chooses to include <osv/app.hh> can use all
it has to offer, and not just the couple of wrapper functions at the end.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170130143247.29697-1-nyh@scylladb.com>