makefile: avoid known warning for exp2l.c
In musl/src/math/exp2l.c we have a known warning:

musl/src/math/exp2l.c:206:22: warning: unused variable ‘v’ [-Wunused-variable]

We already made sure this warning doesn't cause a compilation error, but
let's not show this warning at all.
Instead, we could have fixed the source file to not cause this warning.
But we took this file unmodified from Musl, and don't want to modify it
just to get rid of a warning.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170202135725.6605-1-nyh@scylladb.com>