Really fix wake_lock() race
Commit ad94d79 fixed an important race in wake_lock(): When trying to
wake something in a wait_for(), it is possible that the do_wait_for()
is currently sleeping not on the me->wait(), but rather, after being
woken up for a different reason, on the acquire(mtx). In that case the
wake_lock() needs to allow the thread to continue sleeping for the mutex.

However, while the fix in commit ad94d79 got rid of a crash, a bug
actually remained: The whole point of wake_lock() to cause the given wr
to be eventually marked woken. In other cases, this would because we
queue wr for the mutex, so when the mutex becomes available wr is woken.
But in *this* case (that wait_for_all is waiting for the mutex), wr will
never be woken - a different wr is queued on the mutex (for the same thread).

So we need to wr->clear() here. This will not wake up the thread immediately
(because it is waiting for the mutex, on a different wr), but when it
does wake up, and poll()s all the wr, it will see the woken wr as being
woken, as expected.

Fixes #824

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170221135818.19120-1-nyh@scylladb.com>