Updated ec2-make-ami.py to boto3, added timers and allowed to specify location of img to create AMI from
New ec2-make-ami.py uses boto3 to interact with AWS. Also it logs time it takes to execute
all steps involved in creating an AMI. It also accepts new parameter -i that allows
to specify location of the OSv image file.

Please note that the new script (as old one) needs to be run on EC2 instance that it can
mount EBS volume locally and copy OSv image into /dev/xvdf. Also creating the snapshot takes longest
and is proportinonal to the size of the image.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <1487858081-2694-1-git-send-email-jwkozaczuk@gmail.com>