gdb: fix error on connect on older distros
In commit adf78de4, I noticed that to access values of an enum class,
we needed to add the class's name, which we previously didn't.
I speculated that is only needed with newer gdb. I just confirmed that
indeed, if one tries to debug OSv on an old distro (but still new enough
to compile OSv), the new names do not work and we need to use the old ones
from before commit adf78de4 - otherwise we get errors on "connect".

So in this patch we try both names, trying the new names first, and if
they fail, the old names.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170315184400.25885-1-nyh@scylladb.com>