gdb: fix "osv mmap" on connect on older distros
In commit ce177a50, I fixed "osv mmap" on newer distros as gdb started
requiring quotes around the typename mmu::vma. Unfortunately, these quotes
are considered a syntax error on older gdb. So in this patch we try both
syntaxes, trying the new one first, and if they it fails, the old one.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170316083910.27964-1-nyh@scylladb.com>