osv: xen: split xenbus into generic & PCI parts
Xen platform PCI is x86-specific feature, not currently present on aarch64.

For xenbus to work on both platforms the generic xenbus driver should be probed
directly (under aarch64 in case Xen is detected), but in case of x64 special PCI
driver checks for Xen platform PCI device and probes generic xenbus driver
if such device is found, pretty much like the way it is done in Linux.

Signed-off-by: Sergiy Kibrik <sergiy.kibrik@globallogic.com>
Message-Id: <1490018087-16439-1-git-send-email-sergiy.kibrik@globallogic.com>