math: disable optimizations on files which gcc 4.8.4 miscompiles
It appears (see details in issue #867) that one of the optimizations in
Gcc 4.8.4's "-O1" miscompiles code which uses a function called "round()"
to call "lround()" instead. This causes our lround() function (which
calls our round()) to go into a silly infinite loop.

Sadly, it seems that none of the specific "-fno-*" options to disable
specific optimizations avoid this optimization. So the workaround in this
patch is to use -O0 on the affected source files - those which implement
the various {l,ll}round{,f,l} functions. This will slow these functions
by a little bit, but hopefully they are not heavily used by performance-
minded applications, anyway.

Fixes #867.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170328203039.18789-1-nyh@scylladb.com>