reimplement pthread_barrier_*()
Our pthread barrier implementation had a race described in issue #857,
where one of the threads may start its next round of waiting before the
counter was reset, which can lead to a hang as not enough threads decrease
the counter.

This patch provides a completely new implementation, inspired by
the first algorithm in https://locklessinc.com/articles/barriers/.
Basically, all threads need to agree to reset the counter by bringing
it up from a negative value back to 0 (see more explanations in the
code).

Fixes #857.

The new implementation doesn't use "latch". Rather, we use only a single
mutex, and a single condition variable (actually, a waitqueue which is
like a condition variable but does not require an additional internal
mutex in its implementation).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170409224714.27261-1-nyh@scylladb.com>