Fix SMP initialization of paravirtual KVM clock
Issue #869 reported that the bug which was supposedly fixed by commit
34232de9 has returned when running OSv in a Linux 4.10 host. This patch
fixes the same issue yet again, more correctly, and fixes #869:

For the paravirtual KVM (or Xen) clock, our implementation of uptime()
initially returns 0, until the time that _smp_init is set. Afterwards,
uptime() returns:

     system_time() - _boot_systemtime

The thinking was that when _smp_init is set, so is _boot_systemtime.
That is indeed correct, but the problem is that at this point, system_time()
doesn't necessarily work! The first CPU to call pv_based_clock::setup_cpu()
will set up _boot_systemtime and _smp_init, but at that point other CPUs who
have not yet called setup_cpu() will have a zero system_time(), and as
a result the subtraction above will be negative. So uptime() will return
a negative time on those CPUs, after having previously returned time 0.
Our timer implementation can detect the clock going backward in this manner,
and cause a crash, as reported in issue #869 (and in a long time ago in
commit 34232de9).

The fix is to to set _smp_init not when the first CPU's clock is initialized
(as we did until now), but only only after all CPUs' clocks have been set up -
at which point both system_time() and _boot_systemtime will be correct on
all CPUs.

The downside of this fix is that our clock continues to be stuck at time 0
for even longer during early boot - in my benchmark this time can grow to
almost 0.3 seconds for 32 vcpus. However, this "stuck time" does not cause
any real problems as code which needs functional clocks and preemptive
multitasking will run later anyway.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170418122956.7702-1-nyh@scylladb.com>
Reviewed-by: Benoît Canet <benoit@scylladb.com>