Do not include libosv.so in Java module
This patch removes the unused libosv.so library from the Java module.
Because libosv.so is built based on the kernel executable, when the
executable itself needs to contain libosv.so, we have a vicious circle.
So before this patch, the command
	make clean
	scripts/build image=java fs=ramfs
(note the fs=ramfs, causing all the files to be included in the kernel
executable)
fails with
	strip: 'libosv.so': No such file

This patch simply removes libosv.so from the java module. Nothing seems
to rely on it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170518133747.15842-1-nyh@scylladb.com>