build: allow easily increasing the maximum allowed kernel size
As explained in issue #872, before this patch, the uncompressed OSv kernel
was required to be under 23,068,672 bytes in size. Usually this is not a
problem, but it becomes a problem in certain unoptimized builds, and easily
becomes a problem when the kernel includes a large ram disk. For example,
the command

    scripts/build image=cli,java fs=ramfs

fails before this patch, stating that the kernel is too big. The failure
messages suggests to the user how to increase this limit, but the suggested
workaround is complex, and incomplete.

So this patch adds a simple workaround - a variable "lzkernel_base" in
Makefile which can be increased to raise the limit on the kernel's size.
Now, if the kernel is too big, the Makefile directs the user to increase
the lzkernel_base value and recompile.

lzkernel_base determines where the uncompressed kernel is loaded, and
then decompressed into address kernel_base. So (lzkernel_base - kernel_base)
is the maximum allowed length of the uncompressed kernel.

For this, we had to replace the hard-coded limits and addresses in boot16.S
(which needs to know where the compressed kernel is to have it uncompress
itself), in lzloader.ld (which builds the compressed kernel), in lzloader.cc
(which arbitrarily stopped uncompressing after reading a certain length),
and in Makefile (which verifies that the kernel isn't too large).

Fixes #872.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170518144134.481-1-nyh@scylladb.com>