Fix dladdr() return value
The dladdr() function is supposed to return 0 if no shared library was
found containing the given address, and we neglected to do this and always
returned 1. So this patch fixes the return value of dladdr() to match the
Linux one, and return 0 when it needs to.

One consequence of this bug was that PR_GetLibraryFilePathname() from the
nspr4 library assumed that when dladdr() returned 1, it also returned a
non-null filename, and it could strlen() it. Our wrong return value broke
this assumption.

This patch also fixes the underlying OSv method object::lookup_addr()
to set the returned fname and base to the object's values when the address
belongs to this object - even if not to any known symbol, because we need
this behavior in dladdr().

Fixes #878, but openjdk8-fedora+tomcat still does not work, because
of yet another bug.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170608115157.5552-1-nyh@scylladb.com>