stdio: fix fread() from block device
@justinc1 discovered that strangely, our fread() implementation fails to
read from a block device. This is because even if we fread() 1024 bytes,
the implementation only asks from the readv() system call 1023 bytes,
and then the last byte. The block device then refuses to make the unaligned
read, crashing with an assertion failure.

Turns out that, Musl did this strange thing to work around a bug in Linux's
terminal IO, which is explained in:
http://git.musl-libc.org/cgit/musl/commit/?id=2cff36a84f268c09f4c9dc5a1340652c8e298dc0
OSv does not have this bug - LineDiscipline::read() only blocks once for
the entire readv() scattered buffer, as expected.
So we don't need this workaround, which can only hurt performance, and in
the present case, prevent fread() from working on a block device.

Fixes #877

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170613122824.6539-1-nyh@scylladb.com>