Support the TCP_CORK socket option
This patch adds missing support for the TCP_CORK socket option.
Doing this is easy, because it's already supported by the FreeBSD-based
stack, under the name TCP_NOPUSH (and the different numeric value), and
we just need to do the appropriate translation.

This patch also includes a test that the TCP_CORK works as expected -
we test two consecutive small writes to a socket. Without TCP_CORK, only
the first one happens immediately, the second one is delayed by as much
as 100ms (this is an interaction between the "Nagle's algorithm" in the
sender and "delayed ack" in the reciever). When TCP_CORK is used and
released after the second write, it gets sent immediately.

We saw these 100ms "hangs" of TCP sessions in real benchmarks of nginx,
because its usage of TCP_CORK was silently ignored.

Fixes #883.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>