Return correct errno from socket() on AF_INET6
OSv does not support IPv6 so trying to create a socket() for the address
family AF_INET6 should fail. But it currently fails with EPROTONOSUPPORT,
which is unexpected - considering that the address family itself is
unsupported, it is more logical to return EAFNOSUPPORT. Some existing
software like redis even assume this when testing for IPv6 support.

So in this patch socket() differentiates between an unknown address family
(returning EAFNOSUPPORT) and a known address family but unknown socket type
or protocol (returning EPROTONOSUPPORT, as before).

Fixes #865.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170706215953.16746-1-nyh@scylladb.com>