Support mmap() flag MAP_32BIT
Support the mmap() flag MAP_32BIT which asks for the allocated address
to be in the lower 32 bits. Some applications (like wrk and luajit)
use this flag and need it to work as expected.

By default, OSv mappings look for a hole starting at 0x200000000000.
With this patch and MAP_32BIT, we change that to 0x2000000.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170711145137.28618-1-nyh@scylladb.com>