tests: avoid GOLD linker on three problematic tests
Three tests fail if the system's default linker is gold. In OSv's makefile,
we explicitly choose the bfd linker because of gold's shortcomings, but in
the test Makefile we let the build machine's distribution to choose its own
default.

Fixes failures on tst-mmap, tst-elf-permissions and tst-tls.
Fixes #891 (this is more a work-around than a fix, but it will do for now).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>