gcc 7: replacement for "new" for types of special alignment
OSv has various types which have special alignment requirements - e.g.,
some objects are aligned to cache-line size for performance, some objects
are aligned because of processor requirements (e.g., fpu state needs to
be 64-byte aligned) or ABI requirements (stacks need to be 16-byte aligned).

While C++11 correctly aligns these object on the stack, when created on
the heap with "new", it actually ignores the requested alignment!
The reason is that "new" calls "operator new()" which cannot specify the
desired alignment. We're lucky that in practice missing the alignment
didn't matter: In some cases the alignment was just supposed to be a
performance optimization that is not mandatory. In other cases like
the thread object, because it is large, it was already allocated using
entire pages and ended up getting a 4096 byte alignment, more than enough.

But Gcc starting in release 7, warns about using "new" on a type with a
non-standard alignment requirement. What we do in this patch is to create
a template function aligned_new<T>(...) which can be used instead of new,
and uses aligned_alloc() to implement the properly aligned allocation.

When we switch to C++17 in the future, its "new" fully supports any alignment
(and to do that, it has a version of "operator new" taking an alignment -
see http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2012/n3396.htm)
so we will be able to revert this patch, and go back to using regular new.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>